+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/jobs/JobRowFilterSpec.groovy
@@ -2,14 +2,14 @@
 // Licensed under the terms of the Apache license. Please see LICENSE file distributed with this work for terms.
 package com.yahoo.bard.webservice.jobs
 
-import com.yahoo.bard.webservice.async.jobs.stores.ApiJobStoreFilter
-import com.yahoo.bard.webservice.async.jobs.stores.ApiJobStoreFilterOperation
+import com.yahoo.bard.webservice.async.jobs.stores.JobRowFilter
 import com.yahoo.bard.webservice.web.BadFilterException
+import com.yahoo.bard.webservice.web.FilterOperation
 
 import spock.lang.Specification
 import spock.lang.Unroll
 
-class ApiJobStoreFilterSpec extends Specification {
+class JobRowFilterSpec extends Specification {
 
     @Unroll
     def "Good ApiJobStore filter query #jobField-#op#values parses correctly"() {
@@ -17,12 +17,12 @@ class ApiJobStoreFilterSpec extends Specification {
         String query = "$jobField-$op$values"
 
         when:
-        ApiJobStoreFilter apiJobStoreFilter = new ApiJobStoreFilter(query)
+        JobRowFilter jobRowFilter = new JobRowFilter(query)
 
         then:
-        apiJobStoreFilter.jobField?.name == jobField
-        apiJobStoreFilter.operation == ApiJobStoreFilterOperation.valueOf(op)
-        apiJobStoreFilter.values == expected as Set
+        jobRowFilter.jobField?.name == jobField
+        jobRowFilter.operation == FilterOperation.valueOf(op)
+        jobRowFilter.values == expected as Set
 
         where:
         jobField  | op   | values           | expected
@@ -33,7 +33,7 @@ class ApiJobStoreFilterSpec extends Specification {
     @Unroll
     def "Bad ApiJobStore filter query #query throws #exception.simpleName because #reason"() {
         when:
-        new ApiJobStoreFilter(query)
+        new JobRowFilter(query)
 
         then:
         thrown exception