+++ CHANGELOG.md
@@ -13,7 +13,7 @@ Current
 - [Enrich the ApiJobStore interface] (https://github.com/yahoo/fili/pull/23)
     * `ApiJobStore` Interface now supports filtering `JobRows` in the store
     * Added support for filtering JobRows in `HashJobStore`
-    * Added `ApiJobStoreFilter` to hold filter information
+    * Added `JobRowFilter` to hold filter information
 
 -  [Lookup Dimension Serializer]()
     * Created `LookupDimensionToDimensionSpec` serializer for `LookupDimension`